refactor: Moved the mock_gateway_bot function into Fixtures
