+++ test/support/test_utils.ex
@@ -1,8 +1,6 @@
 defmodule Mobius.TestUtils do
   @moduledoc false
 
-  alias Mobius.Rest.Client
-
   @doc """
   Returns runtime of a function call in milliseconds
 
@@ -19,18 +17,5 @@ defmodule Mobius.TestUtils do
       |> elem(0)
       |> :erlang.convert_time_unit(:microsecond, :millisecond)
     end
-  end
-
-  @doc "Mocks the /gateway/bot request"
-  @spec mock_gateway_bot(integer, integer) :: any
-  def mock_gateway_bot(remaining \\ 1000, reset_after \\ 0) do
-    app_info = %{
-      "shards" => 1,
-      "url" => "wss://gateway.discord.gg",
-      "session_start_limit" => %{"remaining" => remaining, "reset_after" => reset_after}
-    }
-
-    url = Client.base_url() <> "/gateway/bot"
-    Tesla.Mock.mock_global(fn %{url: ^url, method: :get} -> Mobius.Fixtures.json(app_info) end)
   end
 end
