+++ test/test_helper.exs
@@ -1,5 +1,5 @@
 # Mock the API response before Mobius is started
-Mobius.TestUtils.mock_gateway_bot()
+Mobius.Fixtures.mock_gateway_bot()
 
 Application.ensure_all_started(:mobius)
